requirements.txt
pytest>=8.2.0

pandas==2.2.2
numpy==2.0.1
pydantic==2.8.2
python-dotenv==1.0.1

openai==1.40.6
httpx<1.0

streamlit==1.37.1
plotly==5.23.0

rich==13.7.1


